Add integer vector and matrix type support
- Add Numeric constraint (all int/uint/float types) alongside existing Float
- Convert vector/matrix types from array aliases to structs (named fields for
  vectors, Data array for matrices)
- Generate concrete integer types: {I,U}{8,16,32,64}{Vec,Mat}{2,3,4} (48 types)
- Integer Normalize()/Len()/Dist()/Lerp() return float32/Vec types automatically
- Keep generic TVec/TMat with Numeric constraint and limited method set
- Update all tests for struct field access syntax

+++ util.go
@@ -2,7 +2,7 @@ package math
 
 import "math"
 
-func abs[T Float](x T) T {
+func abs[T Numeric](x T) T {
 	if x < 0 {
 		return -x
 	}
@@ -26,7 +26,7 @@ func tan[T Float](x T) T {
 }
 
 // Clamp returns x clamped to [min, max].
-func Clamp[T Float](x, min, max T) T {
+func Clamp[T Numeric](x, min, max T) T {
 	if x < min {
 		return min
 	}
